created graphql file
