+++ src/Components/RecipeBuilderPieces/SubmitButton.jsx
@@ -1,64 +1,12 @@
 import {React} from 'react';
 import {Button,Box} from '@mui/material';
-import { gql } from 'graphql-tag';
 import { useMutation } from '@apollo/client';
-
-const CreateRecipes = gql`
-mutation createRecipe(
-    $title: String!, 
-    $ingredients: String!, 
-    $instructions: String!, 
-    $mealType: String!,
-    $description: String!,
-    $slug: String,
-    $recipePrivacy: String!
-    $prepTime: Int!
-    $cookTime: Int!
-    $servingSize: Int!
-  ) 
-  {
-  createRecipe(
-    data:{
-    title: $title, 
-    ingredients: $ingredients, 
-    instructions: $instructions, 
-    mealType: $mealType,
-    description: $description,
-    slug: $slug,
-    recipePrivacy: $recipePrivacy,
-    cookTime: $cookTime,
-    prepTime: $prepTime,
-    servingSize: $servingSize,
-    }
-  ) 
-  {
-      title
-      ingredients
-      instructions
-      mealType
-      description
-      id
-      recipePrivacy
-      cookTime
-      prepTime
-      servingSize
-  }
-  }
-`
-const PublishRecipes = gql`
-mutation publishRecipe($id: ID){
-  publishRecipe(
-    where: { id: $id }
-  ) {
-    id
-  }
-}
-`
+import {CREATE_MUTATION, PUBLISH_MUTATION} from '../../graphQLQuery';
 
 
 export default function SubmitButton({title, mealType, description, ingredients, instructions, slug, privacy, cookTime, prepTime, servingSize}){
-  const [addRecipes, { loading, error, data }] = useMutation(CreateRecipes);
-  const [publishRecipe] = useMutation(PublishRecipes);
+  const [addRecipes, { loading, error, data }] = useMutation(CREATE_MUTATION);
+  const [publishRecipe] = useMutation(PUBLISH_MUTATION);
 
   if (loading) return <p>Loading...</p>;
   if (error) return console.log(error);
